refactor: :art: Added name safety when exctracting values from form
Because the way of extracting the values is by typing the id of the inputField, is easy to make a mistake and all the program fail. Aldo updating was difficult, because the id has to be searched accross the project. Now we use FormFields and everything will continue be ok